change unsigned int to size_t (prevent warnings)

+++ contracts/musig/bigint/eos_mem_wrapper.cpp
@@ -6,12 +6,12 @@
 using namespace eosio;
 
 // new/delete operator overloading for eos smart contract by mithrilcoin.io
-void *operator new(unsigned int size)
+void *operator new(size_t size)
 {
     return malloc(size);
 }
 
-void *operator new[](unsigned int size)
+void *operator new[](size_t size)
 {
     return malloc(size);
 }
