+++ contracts/musig/bigint/eos_mem_wrapper.hpp
@@ -8,8 +8,8 @@
 
 #include <eosiolib/memory.h>
 
-void *operator new(unsigned int size);
-void *operator new[](unsigned int size);
+void *operator new(size_t size);
+void *operator new[](size_t size);
 void operator delete(void *ptr);
 void operator delete[](void *ptr);
 #endif
